Introducing A Centroid/Converge/Rendezvous/Meet API (#2734)
* optionally allow use of 7bit path_index to track multiple paths in the same edgeset

* pair back edgelabels restriction index to a reasonable max of 254

* use 7 spare bits in edge label to mark the label as belonging to a specific expansion. this allows us to use the same label set for multiple simultaneous path expansions

* first sketch of the primary algorithm. not tested

* initialize best connection and reverse the path, squash me with the previous commit

* halve the cost on the last edge assuming we'll meet in the middle. squash this with the 2 previous commits

* add the API layer and squash some simple mistakes with bit twiddling and typos. it works!

* fix summary info to track the right route

* more TODO notes

* cleanup, dont need to store opposing id in potential convergence point

* changelog

* clang tidy linting

* linting mistakes

* refactor dijkstras interface to specify expansion type. support all types in centroid

* lint

* properly merged master this time

* lint

* add service limits for centroid service. merge all unit test config building into single function making config changes less of a hassle

* add centroid to actor interface and python bindings

* refactor gurka API helper functions to be more generic

* add minimal centroid unit test

* lint

+++ src/sif/motorscootercost.cc
@@ -200,7 +200,7 @@ public:
                        const baldr::GraphId& edgeid,
                        const uint64_t current_time,
                        const uint32_t tz_index,
-                       int& restriction_idx) const override;
+                       uint8_t& restriction_idx) const override;
 
   /**
    * Checks if access is allowed for an edge on the reverse path
@@ -227,7 +227,7 @@ public:
                               const baldr::GraphId& opp_edgeid,
                               const uint64_t current_time,
                               const uint32_t tz_index,
-                              int& restriction_idx) const override;
+                              uint8_t& restriction_idx) const override;
 
   /**
    * Only transit costings are valid for this method call, hence we throw
@@ -374,7 +374,7 @@ bool MotorScooterCost::Allowed(const baldr::DirectedEdge* edge,
                                const baldr::GraphId& edgeid,
                                const uint64_t current_time,
                                const uint32_t tz_index,
-                               int& restriction_idx) const {
+                               uint8_t& restriction_idx) const {
   // Check access, U-turn, and simple turn restriction.
   // Allow U-turns at dead-end nodes.
   if (!IsAccessible(edge) || (!pred.deadend() && pred.opp_local_idx() == edge->localedgeidx()) ||
@@ -397,7 +397,7 @@ bool MotorScooterCost::AllowedReverse(const baldr::DirectedEdge* edge,
                                       const baldr::GraphId& opp_edgeid,
                                       const uint64_t current_time,
                                       const uint32_t tz_index,
-                                      int& restriction_idx) const {
+                                      uint8_t& restriction_idx) const {
   // Check access, U-turn, and simple turn restriction.
   // Allow U-turns at dead-end nodes.
   if (!IsAccessible(opp_edge) || (!pred.deadend() && pred.opp_local_idx() == edge->localedgeidx()) ||
